requirements.txt
requests
httpretty
flask
flask-login
flask-sqlalchemy
sqlalchemy-migrate
flask-whooshalchemy
flask-wtf
flask-babel
guess_language
flipflop
coverage
pylint
autopep8
psycopg2
markdown2
flask-cors
autopep8
coverage